Add files via upload
This is the updated version of the base. The collision is now smooth and does not have the previous "jumpyness". The core loop is also much less large, and is much more readable. This is a significant overall improvement from the original base.

__NEW FEATURES__
+The screen size can be changed in game, just hit the full screen button at the top right, or drag to whatever window size you like.
+The tile sizes may be changed without issues, note to make sure that the size of the tiles are divisible by 2 for best results.
+The size of the player may be changed, note that the player size must be divisible by 4 for best results.

__NOTES__
The above recommendations for sizes are present due to the way which the core loop interacts with collision. Basically, the game draws a one pixel wide line around the unwalkable squares, and assures that the player does not move past this point. This math is done by pixels, and therefore must result in integers. The half size of the player character is often compared against when checking which side the player should be colliding on. Hence the reason why the player must be divisible by 4 in height and width.